requirements.txt
# TODO It is probably not the minimal possible list
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
biopandas==0.5.0.dev0
biopython==1.81
bioservices==1.11.2
bleach==6.0.0
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
dill==0.3.6
easydev==0.12.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.39.4
fqdn==1.5.1
gevent==22.10.2
graphein==1.7.5
greenlet==2.0.2
grequests==0.6.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.15.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jaxtyping==0.2.19
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
loguru==0.7.0
looseversion==1.1.2
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mmtf-python==1.1.3
msgpack==1.0.5
multipledispatch==0.6.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
ordered-set==4.1.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.5.1
plotly==5.14.1
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-cache==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rich-click==1.6.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
suds-community==1.1.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==4.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
url-normalize==1.4.3
urllib3==2.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
wget==3.2
wheel==0.38.4
widgetsnbextension==4.0.7
wrapt==1.15.0
xarray==2023.5.0
xmltodict==0.13.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
